Split the module march and its submodule march.gas into two files (ref issue #197)
